+++ beacon_node/http_api/src/beacon/pool.rs
@@ -10,7 +10,6 @@ use crate::version::{
 use crate::{sync_committees, utils};
 use beacon_chain::observed_operations::ObservationOutcome;
 use beacon_chain::payload_attestation_verification::Error as PayloadAttestationError;
-use beacon_chain::proposer_preferences_verification::ProposerPreferencesError;
 use beacon_chain::{BeaconChain, BeaconChainTypes};
 use bytes::Bytes;
 use eth2::types::{AttestationPoolQuery, EndpointVersion, Failure, GenericResponse};
@@ -25,8 +24,8 @@ use tokio::sync::mpsc::UnboundedSender;
 use tracing::{debug, error, info, warn};
 use types::{
     Attestation, AttestationData, AttesterSlashing, ForkName, PayloadAttestationMessage,
-    ProposerSlashing, SignedBlsToExecutionChange, SignedProposerPreferences, SignedVoluntaryExit,
-    SingleAttestation, SyncCommitteeMessage,
+    ProposerSlashing, SignedBlsToExecutionChange, SignedVoluntaryExit, SingleAttestation,
+    SyncCommitteeMessage,
 };
 use warp::filters::BoxedFilter;
 use warp::{Filter, Reply};
@@ -658,127 +657,6 @@ fn publish_payload_attestation_messages<T: BeaconChainTypes>(
     } else {
         Err(warp_utils::reject::indexed_bad_request(
             "error processing payload attestations".to_string(),
-            failures,
-        ))
-    }
-}
-
-/// POST beacon/pool/proposer_preferences (JSON)
-pub fn post_beacon_pool_proposer_preferences<T: BeaconChainTypes>(
-    network_tx_filter: &NetworkTxFilter<T>,
-    optional_consensus_version_header_filter: OptionalConsensusVersionHeaderFilter,
-    beacon_pool_path: &BeaconPoolPathFilter<T>,
-) -> ResponseFilter {
-    beacon_pool_path
-        .clone()
-        .and(warp::path("proposer_preferences"))
-        .and(warp::path::end())
-        .and(warp_utils::json::json())
-        .and(optional_consensus_version_header_filter)
-        .and(network_tx_filter.clone())
-        .then(
-            |task_spawner: TaskSpawner<T::EthSpec>,
-             chain: Arc<BeaconChain<T>>,
-             preferences: Vec<SignedProposerPreferences>,
-             _fork_name: Option<ForkName>,
-             network_tx: UnboundedSender<NetworkMessage<T::EthSpec>>| {
-                task_spawner.blocking_json_task(Priority::P0, move || {
-                    publish_proposer_preferences(&chain, &network_tx, preferences)
-                })
-            },
-        )
-        .boxed()
-}
-
-/// POST beacon/pool/proposer_preferences (SSZ)
-pub fn post_beacon_pool_proposer_preferences_ssz<T: BeaconChainTypes>(
-    eth_v1: EthV1Filter,
-    task_spawner_filter: TaskSpawnerFilter<T>,
-    chain_filter: ChainFilter<T>,
-    network_tx_filter: NetworkTxFilter<T>,
-) -> ResponseFilter {
-    eth_v1
-        .and(warp::path("beacon"))
-        .and(warp::path("pool"))
-        .and(warp::path("proposer_preferences"))
-        .and(warp::path::end())
-        .and(warp::body::bytes())
-        .and(task_spawner_filter)
-        .and(chain_filter)
-        .and(network_tx_filter)
-        .then(
-            |body_bytes: Bytes,
-             task_spawner: TaskSpawner<T::EthSpec>,
-             chain: Arc<BeaconChain<T>>,
-             network_tx: UnboundedSender<NetworkMessage<T::EthSpec>>| {
-                task_spawner.blocking_json_task(Priority::P0, move || {
-                    let item_len = <SignedProposerPreferences as Encode>::ssz_fixed_len();
-                    if !body_bytes.len().is_multiple_of(item_len) {
-                        return Err(warp_utils::reject::custom_bad_request(format!(
-                            "SSZ body length {} is not a multiple of SignedProposerPreferences size {}",
-                            body_bytes.len(),
-                            item_len,
-                        )));
-                    }
-                    let preferences: Vec<SignedProposerPreferences> = body_bytes
-                        .chunks(item_len)
-                        .map(|chunk| {
-                            SignedProposerPreferences::from_ssz_bytes(chunk).map_err(|e| {
-                                warp_utils::reject::custom_bad_request(format!(
-                                    "invalid SSZ: {e:?}"
-                                ))
-                            })
-                        })
-                        .collect::<Result<_, _>>()?;
-                    publish_proposer_preferences(&chain, &network_tx, preferences)
-                })
-            },
-        )
-        .boxed()
-}
-
-fn publish_proposer_preferences<T: BeaconChainTypes>(
-    chain: &BeaconChain<T>,
-    network_tx: &UnboundedSender<NetworkMessage<T::EthSpec>>,
-    preferences_list: Vec<SignedProposerPreferences>,
-) -> Result<(), warp::Rejection> {
-    let mut failures = vec![];
-    let mut num_already_known = 0;
-
-    for (index, preferences) in preferences_list.into_iter().enumerate() {
-        match chain.verify_proposer_preferences_for_gossip(Arc::new(preferences.clone())) {
-            Ok(_verified) => {
-                utils::publish_pubsub_message(
-                    network_tx,
-                    PubsubMessage::ProposerPreferences(Box::new(preferences)),
-                )?;
-            }
-            Err(ProposerPreferencesError::AlreadySeen { .. }) => {
-                num_already_known += 1;
-            }
-            Err(e) => {
-                error!(
-                    error = ?e,
-                    request_index = index,
-                    "Failure verifying proposer preferences for gossip"
-                );
-                failures.push(Failure::new(index, format!("{e:?}")));
-            }
-        }
-    }
-
-    if num_already_known > 0 {
-        debug!(
-            count = num_already_known,
-            "Some proposer preferences already known"
-        );
-    }
-
-    if failures.is_empty() {
-        Ok(())
-    } else {
-        Err(warp_utils::reject::indexed_bad_request(
-            "error processing proposer preferences".to_string(),
             failures,
         ))
     }
