+++ beacon_node/http_api/src/lib.rs
@@ -1510,15 +1510,16 @@ pub fn serve<T: BeaconChainTypes>(
     let post_beacon_pool_bls_to_execution_changes =
         post_beacon_pool_bls_to_execution_changes(&network_tx_filter, &beacon_pool_path);
 
-    // POST beacon/pool/proposer_preferences (JSON)
-    let post_beacon_pool_proposer_preferences = post_beacon_pool_proposer_preferences(
-        &network_tx_filter,
-        optional_consensus_version_header_filter,
-        &beacon_pool_path,
+    // POST validator/proposer_preferences (JSON)
+    let post_validator_proposer_preferences = post_validator_proposer_preferences(
+        eth_v1.clone(),
+        task_spawner_filter.clone(),
+        chain_filter.clone(),
+        network_tx_filter.clone(),
     );
 
-    // POST beacon/pool/proposer_preferences (SSZ)
-    let post_beacon_pool_proposer_preferences_ssz = post_beacon_pool_proposer_preferences_ssz(
+    // POST validator/proposer_preferences (SSZ)
+    let post_validator_proposer_preferences_ssz = post_validator_proposer_preferences_ssz(
         eth_v1.clone(),
         task_spawner_filter.clone(),
         chain_filter.clone(),
@@ -3432,7 +3433,7 @@ pub fn serve<T: BeaconChainTypes>(
                             .uor(post_beacon_blinded_blocks_v2_ssz)
                             .uor(post_beacon_execution_payload_envelope_ssz)
                             .uor(post_beacon_pool_payload_attestations_ssz)
-                            .uor(post_beacon_pool_proposer_preferences_ssz),
+                            .uor(post_validator_proposer_preferences_ssz),
                     )
                     .uor(post_beacon_blocks)
                     .uor(post_beacon_blinded_blocks)
@@ -3445,7 +3446,7 @@ pub fn serve<T: BeaconChainTypes>(
                     .uor(post_beacon_pool_sync_committees)
                     .uor(post_beacon_pool_payload_attestations)
                     .uor(post_beacon_pool_bls_to_execution_changes)
-                    .uor(post_beacon_pool_proposer_preferences)
+                    .uor(post_validator_proposer_preferences)
                     .uor(post_beacon_execution_payload_envelope)
                     .uor(post_beacon_state_validators)
                     .uor(post_beacon_state_validator_balances)
